#38 Removed CreateTransactionForm
- Removed CreateTransactionForm, now replaced with CreateTxnForm

- CreateTxnForm now has an Account field
- It can receive either an Account or List<Account>
- It now updates baseCurrency and targetCurrency correctly when chosenAccount changes
- Moved CreateTransaction(Provider) into create_txn_form_widget.dart
