+++ lib/presentation/debug/debug_transactions/providers/debug_transactions_provider.dart
@@ -11,88 +11,6 @@ import 'package:riverpod_annotation/riverpod_annotation.dart';
 part 'debug_transactions_provider.g.dart';
 
 @riverpod
-class CreateTransaction extends _$CreateTransaction {
-  @override
-  Future<Option<Unit>> build() async {
-    return const Option.none();
-  }
-
-  Future<void> createTransaction({
-    required String name,
-    required String? description,
-    required String accountId,
-    required TransactionType transactionType,
-    required double? baseAmount,
-    required String baseCurrency,
-    required double? exchangeRate,
-    required String? targetCurrency,
-    required double? targetAmount,
-  }) async {
-    state = const AsyncValue.loading();
-
-    debugPrint('retrieving logged in user...');
-    final failureOrPecuniaUser = await ref.read(authRepoProvider).getLoggedInUser().run();
-
-    final pecuniaUser = failureOrPecuniaUser.fold(
-      (failure) {
-        state = AsyncValue.error(failure, failure.stackTrace);
-      },
-      (pecuniaUser) => pecuniaUser,
-    );
-
-    double actualBaseAmount;
-
-    // ! Temporary Workaround. Remove this when we have a proper way to handle this.
-    if (baseAmount == null) {
-      actualBaseAmount = targetAmount! / exchangeRate!;
-    } else {
-      actualBaseAmount = baseAmount;
-    }
-
-    if (pecuniaUser != null) {
-      debugPrint('user exists, creating transaction...');
-      (await ref
-              .read(transactionsRepoProvider)
-              .createTransaction(
-                name: name,
-                creatorUid: pecuniaUser.uid,
-                transactionDate: DateTime.now(),
-                accountId: accountId,
-                type: transactionType.typeAsString,
-                baseAmount: actualBaseAmount,
-                baseCurrency: baseCurrency,
-                transactionDescription: description,
-                exchangeRate: exchangeRate,
-                targetCurrency: targetCurrency,
-                targetAmount: targetAmount,
-              )
-              .run())
-          .fold(
-        (l) => state = AsyncError(l, l.stackTrace),
-        (r) {
-          ref.invalidate(getAllTransactionsProvider);
-          state = AsyncData(Option.of(r));
-        },
-      );
-      debugPrint('''
-      name: $name,
-      creatorUid: ${pecuniaUser.uid},
-      transactionDate: ${DateTime.now()},
-      accountId: $accountId,
-      type: $transactionType,
-      baseAmount: $baseAmount,
-      actualBaseAmount: $actualBaseAmount,
-      baseCurrency: $baseCurrency,
-      transactionDescription: $description,
-      exchangeRate: $exchangeRate,
-      targetCurrency: $targetCurrency,
-      targetAmount: $targetAmount,
-''');
-    }
-  }
-}
-
-@riverpod
 class DeleteTransaction extends _$DeleteTransaction {
   @override
   Future<Option<Unit>> build() async {
